backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-filter>=23.1
djoser>=2.1
psycopg2-binary>=2.9
Pillow>=9.5
gunicorn>=20.1
